Implement the Refresh & Submit buttons
