+++ src/index.html
@@ -8,6 +8,7 @@
   </head>
   <body>
     <h1 class="header">Leaderboard</h1>
+    <p class="alert"></p>
     <main class="project">
       <section class="board">
         <div class="board-header">
